+++ test/compute-engine/latex-syntax/trigonometry.test.ts
@@ -26,25 +26,25 @@ describe('TRIGONOMETRIC FUNCTIONS inverse, prime', () => {
     expect(check("\\sin^{-1}'(x)")).toMatchInlineSnapshot(`
       box       = ["Apply", ["Derivative", ["InverseFunction", "Sin"]], "x"]
       canonical = ["Apply", ["Derivative", "Arcsin"], "x"]
-      eval-auto = (-x^2 + 1)^(-1/2)
+      eval-auto = 1 / sqrt(1 - x^2)
     `));
   test(`\\sin^{-1}''(x)`, () =>
     expect(check("\\sin^{-1}''(x)")).toMatchInlineSnapshot(`
       box       = ["Apply", ["Derivative", ["InverseFunction", "Sin"], 2], "x"]
       canonical = ["Apply", ["Derivative", "Arcsin", 2], "x"]
-      eval-auto = x / (-x^2 + 1)^(3/2)
+      eval-auto = x / (1 - x^2)^(3/2)
     `));
   test(`\\cos^{-1\\doubleprime}(x)`, () =>
     expect(check('\\cos^{-1\\doubleprime}(x)')).toMatchInlineSnapshot(`
       box       = ["Apply", ["Derivative", ["InverseFunction", "Cos"], 2], "x"]
       canonical = ["Apply", ["Derivative", "Arccos", 2], "x"]
-      eval-auto = -x / (-x^2 + 1)^(3/2)
+      eval-auto = -x / (1 - x^2)^(3/2)
     `));
   test(`\\cos^{-1}\\doubleprime(x)`, () =>
     expect(check('\\cos^{-1}\\doubleprime(x)')).toMatchInlineSnapshot(`
       box       = ["Apply", ["Derivative", ["InverseFunction", "Cos"], 2], "x"]
       canonical = ["Apply", ["Derivative", "Arccos", 2], "x"]
-      eval-auto = -x / (-x^2 + 1)^(3/2)
+      eval-auto = -x / (1 - x^2)^(3/2)
     `));
 });